+++ pydantic_ai_harness/__init__.py
@@ -4,19 +4,15 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from .code_mode import CodeMode
-    from .dynamic_workflow import DynamicWorkflow, WorkflowAgent, WorkflowResourceLimits
     from .filesystem import FileSystem
     from .logfire import ManagedPrompt
     from .shell import Shell
 
 __all__ = [
     'CodeMode',
-    'DynamicWorkflow',
     'FileSystem',
     'ManagedPrompt',
     'Shell',
-    'WorkflowAgent',
-    'WorkflowResourceLimits',
 ]
 
 
@@ -25,18 +21,6 @@ def __getattr__(name: str) -> object:
         from .code_mode import CodeMode
 
         return CodeMode
-    if name == 'DynamicWorkflow':
-        from .dynamic_workflow import DynamicWorkflow
-
-        return DynamicWorkflow
-    if name == 'WorkflowAgent':
-        from .dynamic_workflow import WorkflowAgent
-
-        return WorkflowAgent
-    if name == 'WorkflowResourceLimits':  # pragma: no cover
-        from .dynamic_workflow import WorkflowResourceLimits
-
-        return WorkflowResourceLimits
     if name == 'FileSystem':
         from .filesystem import FileSystem
