--- pydantic_ai_harness/dynamic_workflow/__init__.py
@@ -1,10 +1,0 @@
-"""Dynamic workflow capability: orchestrate sub-agents from a sandboxed Python script."""
-
-from pydantic_ai_harness.dynamic_workflow._capability import DynamicWorkflow
-from pydantic_ai_harness.dynamic_workflow._toolset import (
-    DynamicWorkflowToolset,
-    WorkflowAgent,
-    WorkflowResourceLimits,
-)
-
-__all__ = ['DynamicWorkflow', 'DynamicWorkflowToolset', 'WorkflowAgent', 'WorkflowResourceLimits']
