+++ pydantic_ai_harness/experimental/dynamic_workflow/__init__.py
@@ -1,0 +1,13 @@
+"""Dynamic workflow capability: orchestrate sub-agents from a sandboxed Python script."""
+
+from pydantic_ai_harness.experimental._warn import warn_experimental
+from pydantic_ai_harness.experimental.dynamic_workflow._capability import DynamicWorkflow
+from pydantic_ai_harness.experimental.dynamic_workflow._toolset import (
+    DynamicWorkflowToolset,
+    WorkflowAgent,
+    WorkflowResourceLimits,
+)
+
+warn_experimental('dynamic_workflow')
+
+__all__ = ['DynamicWorkflow', 'DynamicWorkflowToolset', 'WorkflowAgent', 'WorkflowResourceLimits']
